+++ src/pandashape/describers/GeneralDescriber.py
@@ -17,13 +17,4 @@ class GeneralDescriber(Describer):
                 f"Columns with one or more null values: {columns_with_nulls}"
             )
 
-        # LE CANDIDATES
-        object_typed_columns = df.select_dtypes(
-            include='object').columns.array.to_numpy()
-
-        if len(object_typed_columns) > 0:
-            messages.append(
-                f"Columns of type \"object\" (may need label encoding): {object_typed_columns}"
-            )
-
         return messages