+++ src/pandashape/describers/__init__.py
@@ -1,3 +1,4 @@
+from .CategoricalDescriber import CategoricalDescriber
 from .CorrelationDescriber import CorrelationDescriber
 from .Describer import Describer
 from .DistributionDescriber import DistributionDescriber